Rewrite the room transition routine. There is now a new 'obstacle avoidance inhibition circle' around the room entrance. In this circle, the obstacle avoidance is disabled and the robot runs straight forward. If there is an other robot in the way, just push it, no a single frack is given.
For an unknown reason, the simulation does crap if 'closestRoomAngle' and 'closestRoomDistance' are not initialized at the begining of each step().
TODO: read the brodcasted signals and change the target room accordingly. Beware that the findClosestRoom() function should thus not be used in this case. You read the other's qualities, then you go to the best one broadcasted. If it's yours (that is, if you did not sense any better quality than yours), just go back to your room.
Note: if you think you are in the highest quality room, you should stay there longer.